requirements.txt
boto==2.40.0
dj-database-url==0.4.1
dj-static==0.0.6
Django==1.9.6
django-storages==1.4.1
gunicorn==19.5.0
gviz-api.py==1.8.2
psycopg2==2.6.1
static3==0.7.0
